+++ MOP/src/GameObjects/Items/Items.cs
@@ -100,6 +100,11 @@ namespace MOP
                 {
                     cdEnchancedObjects[i].AddComponent<ItemHook>();
                 }
+
+                // Create shop table chekc, for when the CDs are bought
+                GameObject itemCheck = new GameObject("MopItemAreaCheck");
+                itemCheck.transform.position = GameObject.Find("SpawnItemStore").transform.position;
+                itemCheck.AddComponent<ItemSpawnCheck>();
             }
         }