+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
-## 2.1.1 Development 1 (tba)
+## 2.1.1 Development 2 (tba)
 
 ### Added
 
-- Added support for CD Player Enhaned Mod
+- CD Player Enhanced Mod: added check for when the CDs have been bought
+
+## 2.1.1 Development 1 (02.02.2020)
+
+### Added
+
+- Added support for CD Player Enhanced Mod
 
 ### Changes